#173 move control flows to UIService
